requirements.txt
git+https://github.com/facebookresearch/segment-anything.git
gradio==3.50.2
numpy
omegaconf==2.3.0
einops==0.6.0
transformers==4.27.3
pytorch-lightning==2.0.1.post0
diffusers==0.14.0
diffusers==0.14.0
google-cloud-storage
google-cloud-pubsub
requests
google-auth